chore: move pallas library to sub-dir
