+++ Behaviors/Pallas/Classic/Priest.js
@@ -1,4 +1,4 @@
-import { Behavior } from '/Behavior.js'
+import { Behavior } from '/Core/Behavior.js'
 
 export class PriestBehavior extends Behavior {
 }
