+++ Behaviors/Pallas/Classic/Warlock.js
@@ -1,4 +1,4 @@
-import { Behavior } from '/Behavior.js'
+import { Behavior } from '/Core/Behavior.js'
 
 export class WarlockBehavior extends Behavior {
 }
