+++ Behaviors/Pallas/Retail/Warrior.js
@@ -1,9 +1,9 @@
 import { GameFlavor } from 'wow';
-import { Behavior, BehaviorContext } from '/Behavior.js'
-import * as bt from '/BehaviorTree';
-import { common } from '/Common';
-import { Specialization } from '/Data/Specialization'
-import { spell } from '/Spell';
+import { Behavior, BehaviorContext } from '/Core/Behavior.js'
+import * as bt from '/Core/BehaviorTree';
+import { common } from '/Core/Common';
+import { Specialization } from '/Core/Data/Specialization'
+import { spell } from '/Core/Spell';
 
 export class WarriorFuryBehavior extends Behavior {
   context = BehaviorContext.All;
