+++ Core/ObjectManager.js
