+++ Core/Targeting.js
@@ -1,5 +1,5 @@
 import * as bt from './BehaviorTree';
-import { me } from '/ObjectManager';
+import { me } from './ObjectManager';
 
 export class Targeting {
   static ensureTarget() {
